get sample details when navigating
